fixed data type of eeprom::sfr_address in sfr.hpp to match sfr.cpp

+++ src/sfr.hpp
@@ -336,7 +336,7 @@ namespace sfr {
         extern SFRField<bool> alloted_time_passed;
         extern int sfr_last_write_time;
         extern SFRField<uint32_t> sfr_write_step_time;
-        extern int sfr_address;
+        extern uint16_t sfr_address;
         extern SFRField<uint32_t> sfr_address_age;
     } // namespace eeprom
 };    // namespace sfr
